requirements.txt
Flask
flask-jwt-extended
Flask-sqlalchemy
flask-bcrypt
Flask-form
Flask-WTF
Flask-Mail
selenium
webdriver_manager
flask-marshmallow
python-dotenv
flask_mysqldb
mysql-connector-python
phonenumbers
Bootstrap-Flask